Trizz 202 refactor pages components intoa pages folder (#345)
* My testy file

* Relocated CalendarPage.js

* Relocate LandingPage component

* fix: invalid index.css imports

* chore: Remove unknown file

* lint fixes

---------
