+++ client/src/pages/CalendarPage.js
@@ -1,6 +1,6 @@
 import useDate from "hooks/useDate";
-import Calendar from "../calendar/Calendar";
-import CalendarHeader from "../calendarHeader";
+import Calendar from "../features/calendar/Calendar";
+import CalendarHeader from "../features/calendarHeader";
 import FormProvider from "contexts/FormContext";
 import Modal from "features/modal/Modal";
 import UserForm from "features/form/UserForm";
